% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateO.R
\name{EvaluateO}
\alias{EvaluateO}
\title{Estimate the amount of O2 for a given FeOt and FeO/Fe2O3 mass ratio}
\usage{
EvaluateO(FeOt, FeRatio)
}
\arguments{
\item{FeOt}{amount of FeOt, wt.pct.}

\item{FeRatio}{FeO/Fe2O3 weight ratio (Middlemost 1989)}
}
\description{
Estimate the amount of O2 for a given FeOt and FeO/Fe2O3 mass ratio
}
