% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseBoundaries.R
\name{findLowerStability}
\alias{findLowerStability}
\alias{findSolidus}
\alias{findUpperStability}
\alias{findLiquidus}
\title{Find phase boundaries}
\usage{
findLowerStability(
  phasename,
  Tlow = 550,
  Thigh = 1450,
  maxIter = 32,
  AmountTolerance = 0.5/100,
  verbose = F,
  printPhaseBoundaryResults = T,
  ...
)

findSolidus(
  Tlow = 550,
  Thigh = 1450,
  maxIter = 32,
  AmountTolerance = 0.5/100,
  verbose = F,
  printPhaseBoundaryResults = T,
  ...
)

findUpperStability(
  phasename,
  Tlow = 550,
  Thigh = 1450,
  maxIter = 32,
  AmountTolerance = 0.5/100,
  verbose = F,
  printPhaseBoundaryResults = T,
  ...
)

findLiquidus(
  Tlow = 850,
  Thigh = 1650,
  maxIter = 32,
  AmountTolerance = 0.5/100,
  verbose = F,
  printPhaseBoundaryResults = T,
  ...
)
}
\arguments{
\item{phasename}{(string) the name of the phase to look for, as known from MAGEMin.}

\item{Tlow, Thigh}{Initial search range (see details)}

\item{maxIter}{Maximum number of iterations (see details)}

\item{AmountTolerance}{The value of the phase under which the boundary is considered to be reached}

\item{verbose}{Report information on every step (mostly for debuging purposes)}

\item{printPhaseBoundaryResults}{Report on the outcome of the calculation}

\item{...}{Further arguments that will be passed to \code{\link[=MAGEMin]{MAGEMin()}}}
}
\value{
A list containing the MAGEMin output at the phase boundary, converted to a R list.
This is the same format as in \code{\link[=MAGEMin]{MAGEMin()}} except that 2 fields have been added,
\verb{$lower_<PHASE>_boundary_found} (a boolean) and \verb{$iterations} (the number of iterations)
in the case of \code{findSolidus()} or \code{findLiquidus()} the field is called \verb{$solidus_found} or \verb{$liquidus_found}

@details
These functions use the bissection method to locate the position of a phase boundary, i.e. where
its abundance becomes 0. A lower phase boundary (including solidus) is defined as the lowest temperature at
which a phase is present, an upper phase boundary as the highest. This works well for
phases with simple range (for instance liquid which is never present under the solidus and always above). For phases
that are present only in part of the range (most other minerals), this may become problematic if the initial guess is not correct.

The run time strongly depends on the number of iterations needed. They depend marginally on the original temperature range \verb{[Tlow,Thigh]}
(because the nature of the method is that the range is halved each time, so doubling the initial start range
only adds one iteration) but much more strongly on the resolution required (\code{AmountTolerance}). The max number
of iterations \code{maxIter} is not very likely to be reached (unless you put a very small number)

So the user has to decide on a crude but reasonably quick approximation (high tolerance), or a slower but finer one (low tolerance).

Fixing the initial range is somewhat more tricky. For a simple phase (liquid) there is no reason to keep it too narrow.
For a more complex stability field, you probably need to have an idea of what you are doing.

One of the reason a computation may run for a long time is when the system does not find a solution, and
is "stuck" at one of the bounds (upper/lower). Possible solutions would be (i) to decrease the max number of iterations
or (ii) to widen the bound - remember, this adds only one or two iterations but ensures that you will find
the solution, most likely long before reaching \code{maxIter}.

In case of a strongly eutectic behaviour (lots of melt forming at a given T), the function will struggle
to find the minimum. I cannot think of an esay fix. Probably set a reasonable \code{maxIter} and hope for
the best. Keep in mind that you half the range with each iteration, so with \code{maxIter=10} you divided
the original range by 10 and you are probably within a degree if the actual temperature...

\code{findLiquidus()} follows the same logic, but looks for the highest temperature where liquid is \emph{not}
the only stable phase.
}
\description{
Various functions that allow to find lower and upper stability bounds for phases.
\code{findLowerStability()} and \code{findUpperStability()} can be used on any phase, specified with \code{phasename=...}.
\code{findSolidus()} is a convenience function, strictly equivalent to \code{findLowerStability(phasename="liq",...)}.
\code{findLiquidus()} operates in a similar way.
}
