% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGEMin_Parse.R
\name{getSSComp}
\alias{getSSComp}
\title{Extract the composition of a solid solution phase from MAGEMin results}
\usage{
getSSComp(MMo, phName, unit = "wt")
}
\arguments{
\item{MMo}{A R object with MAGEMin data, probably generated by MAGEMin()}

\item{phName}{The name of the phase to extract}

\item{unit}{c("mol","wt","apfu","emmol","emwt") Unit in which the composition is expressed:
can be moles, weight, atom per formula unit or end-member proportions (in mol or wt).}
}
\value{
A named vector of phase composition
}
\description{
Extract the composition of a solid solution phase from MAGEMin results
}
