% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGEMin_Parse.R
\name{JuliaStructToList}
\alias{JuliaStructToList}
\title{Convert a julia struct to a list}
\usage{
JuliaStructToList(julia_object)
}
\arguments{
\item{julia_object}{A Julia struct (made of several fields)}
}
\value{
A R list with the content of the input object
}
\description{
Ancillary function to convert a complex julia struct to a R list.
Mostly called by \code{\link[=MAGEMin]{MAGEMin()}}, but it could be used to parse sub-items
of the MAGEMin struct.
}
\details{
Simple types (vectors of atomic types, etc) will be converted to
R equivalents, but more complex types will be returned as a JuliaObject
that the user would need to manually process
}
