% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGEMin_connect.R
\name{MAGEMin}
\alias{MAGEMin}
\title{Do the actual MAGEMin calculation. This takes some time, typically 200-500 ms
for one run for MAGEMin proper + a bit of overhead for the various post-processing.
It is maybe clever to wrap it in a progress bar}
\usage{
MAGEMin(
  Xoxides,
  X,
  Pkbar,
  TC,
  extraMAGEMinParams = "name_solvus = true",
  showResults = F
)
}
\arguments{
\item{Xoxides}{Character vector. Name of the oxides to use. They should exist in the MAGEMin database.}

\item{X}{Numeric vector. Composition of the system for each oxide, matching Xoxides}

\item{Pkbar}{numeric. Pressure, in kbar. The function makes sure that an integer is converted to a decimal
(julia makes the difference)}

\item{TC}{numeric. Temperature, in degree celsius}

\item{extraMAGEMinParams}{String. Extra parameters to be passed to MAGEMin minimization,
for instance 'B=1.0' (to pass a buffer offset)}

\item{showResults}{Boolean. If true, print julia's summary of the minimlization.}
}
\value{
A list containing the MAGEMin output, converted to a R list. Most fields of the MAGEMin object
(technically a MAGEMin_C gmin_struct) are straightforwards to convert to R. However, the julia structure contains
complex types for SS_vec (solution phases), mSS_vec (metastable solutions) and PP_vec (pure phases). SS_vec and PP_vec are
also parsed and converted to (sub)lists.
Note that both still contain complex types, that are NOT converted to R and remain as JuliaObjects. In SS_vec
these are the end-members compositions, that are known by definition.
}
\description{
Do the actual MAGEMin calculation. This takes some time, typically 200-500 ms
for one run for MAGEMin proper + a bit of overhead for the various post-processing.
It is maybe clever to wrap it in a progress bar
}
\note{
A side effect of the function is also to create a variable out in julia, that can be accessed
with julia_console() or julia_command() if needed.
julia_command("out") will invoke julia's print method and display summary info.
If the julia object needs to be preserved, the best is probably to do something like
foo <- julia_eval("out") (in R)
}
