% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseBoundaries.R
\name{findSolidus}
\alias{findSolidus}
\alias{findLowerStability}
\title{Find phase boundaries}
\usage{
findSolidus(printPhaseBoundaryResults = T, ...)

findLowerStability(
  phasename,
  Tlow = 550,
  Thigh = 1450,
  maxIter = 32,
  AmountTolerance = 0.5/100,
  verbose = F,
  printPhaseBoundaryResults = T,
  ...
)
}
\arguments{
\item{printPhaseBoundaryResults}{Report on the outcome of the calculation}

\item{...}{Further arguments that will be passed to \code{\link[=MAGEMin]{MAGEMin()}}}

\item{phasename}{(string) the name of the phase to look for, as known from MAGEMin.}

\item{Tlow, Thigh}{Initial search range (see details)}

\item{maxIter}{Maximum number of iterations (see details)}

\item{AmountTolerance}{The value of the phase under which the boundary is considered to be reached}

\item{verbose}{Report information on every step (mostly for debuging purposes)}
}
\value{
A list containing the MAGEMin output at the phase boundary, converted to a R list.
This is the same format as in \code{\link[=MAGEMin]{MAGEMin()}} except that 2 fields have been added,
\verb{$lower_<PHASE>_Boundary_found} (a boolean) and \verb{$iterations} (the number of iterations)
@details
These functions use the bissection method to locate the position of a phase boundary, i.e. where
its abundance becomes 0. A lower phase boundary (including solidus) is defined as the lowest temperature at
which a phase is present, an upper phase boundary as the highest. This works well for
phases with simple range (for instance liquid which is never present under the solidus and always above). For phases
that are present only in part of the range (most other minerals), this may become problematic if the initial guess is not correct.

The run time strongly depends on the number of iterations needed. They depend marginally on the original temperature range \verb{[Tlow,Thigh]}
(because the nature of the method is that the range is halved each time, so doubling the initial start range
only adds one iteration) but much more strongly on the resolution required (\code{AmountTolerance}). The max number
of iterations \code{maxIter} is not very likely to be reached (unless you put a very small number)

So the user has to decide on a crude but reasonably quick approximation (high tolerance), or a slower but finer one (low tolerance).

Fixing the initial range is somewhat more tricky. For a simple phase (liquid) there is no reason to keep it too narrow.
For a more complex stability field, you probably need to have an idea of what you are doing.
}
\description{
Various functions that allow to find lower and upper stability bounds for phases.
\code{findLowerStability()} and \code{findLowerStability()} can be used on any phase, specified with \verb{phasename=``. }findSolidus()\verb{is a convenience function, strictly equivalent to}findLowerStability(phasename="liq")`.
}
