% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGEMin_Parse.R
\name{getPhProp}
\alias{getPhProp}
\title{Get phases propoportions}
\usage{
getPhProp(MMo)
}
\arguments{
\item{MMo}{A R object with MAGEMin data, probably generated by MAGEMin()}
}
\value{
A named vector of phase proportions
}
\description{
Extract the (wt\%) proportions of each stable phase (pure or solid solution)
from a MAGEMin object.
}
