% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGEMin_Parse.R
\name{JuliaStructToList}
\alias{JuliaStructToList}
\title{Ancillary function to convert a complex julia struct to a R list
Simple types (vectors of atomic types, etc) will be converted to 
R equivalents, but more complex types will be returned as a JuliaObject
that the user would need to manually process}
\usage{
JuliaStructToList(julia_object)
}
\arguments{
\item{julia_object}{A Julia struct (made of several fields)}
}
\value{
A R list with the content of the input object
}
\description{
Ancillary function to convert a complex julia struct to a R list
Simple types (vectors of atomic types, etc) will be converted to 
R equivalents, but more complex types will be returned as a JuliaObject
that the user would need to manually process
}
